% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\docType{methods}
\name{addDefaultMed}
\alias{addDefaultMed}
\alias{addDefaultMed,Arena-method}
\title{Add default medium of an organism to arena.}
\usage{
addDefaultMed(object, org)

\S4method{addDefaultMed}{Arena}(object, org)
}
\arguments{
\item{object}{An object of class Arena.}

\item{org}{An object of class Organism}
}
\description{
The generic function \code{addDefaultMed} uses the lower bounds defined in an organism's model file to compose minimal medium.
}

