% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\docType{methods}
\name{addEssentialMed}
\alias{addEssentialMed}
\alias{addEssentialMed,Arena-method}
\title{Add minimal medium of an organism to arena.}
\usage{
addEssentialMed(object, org)

\S4method{addEssentialMed}{Arena}(object, org)
}
\arguments{
\item{object}{An object of class Arena.}

\item{org}{An object of class Organism}
}
\description{
The generic function \code{addEssentialMed} uses flux variability analysis to determine a essential growth medium  components (eg. cofactors)
}

