% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Arena.R
\name{addEval}
\alias{addEval}
\title{Function for adding a simulation step}
\usage{
addEval(object, arena, replace = F)
}
\arguments{
\item{object}{An object of class Eval.}

\item{arena}{An object of class Arena.}

\item{replace}{A boolean variable indicating if the last simulation step should be replaced by the new simulation step \code{arena}.}
}
\description{
The generic function \code{addEval} adds results of a simulation step to an \code{Eval} object.
}
\details{
The function \code{addEval} can be used in iterations to manipulate an \code{Arena} object and store the results in an \code{Eval} object.
}
\examples{
\dontrun{
ecore <- model #get Escherichia coli core metabolic model
bac <- Bac(ecore,deathrate=0.05,duplirate=0.5,
           growthlimit=0.05,growtype="exponential") #initialize a bacterium
arena <- Arena(20,20) #initialize the environment
addOrg(arena,bac,amount=10) #add 10 organisms
addSubs(arena,40) #add all possible substances
eval <- simulate(arena,10)
addEval(eval,arena)
}
}
\seealso{
\code{\link{Eval-class}} and \code{\link{Arena-class}}
}

