% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\docType{methods}
\name{addMinMed}
\alias{addMinMed}
\alias{addMinMed,Arena-method}
\title{Add minimal medium of an organism to arena.}
\usage{
addMinMed(object, org)

\S4method{addMinMed}{Arena}(object, org)
}
\arguments{
\item{object}{An object of class Arena.}

\item{org}{An object of class Organism}
}
\description{
The generic function \code{addMinMed} uses the lower bounds defined in an organism's model file to compose minimal medium.
}

