% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Arena.R
\name{addOrg}
\alias{addOrg}
\title{Add individuals to the environment}
\usage{
addOrg(object, specI, amount, x = NULL, y = NULL, growth = 1, ...)
}
\arguments{
\item{object}{An object of class Arena}

\item{specI}{An object of class Organism}

\item{amount}{Number of organisms to add}

\item{x}{x positions of individuals on the grid}

\item{y}{y positions of individuals on the grid}

\item{growth}{Starting biomass of organisms}
}
\description{
The generic function \code{addOrg} adds individuals to the environment.
}
\details{
## S4 method for signature 'Arena'

The arguments \code{x} and \code{y} should be in the same length as the number of organisms added (given by the argument \code{amount}).
}
\examples{
\dontrun{
ecore <- model
bac <- Bac(ecore,deathrate=0.05,duplirate=0.5,growthlimit=0.05,growtype="exponential")
arena <- Arena(20,20)
addOrg(arena, bac, amount=10)
}
}
\seealso{
\code{\link{Arena-class}} and \code{\link{Bac-class}}
}

