% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Arena.R
\name{addSubs}
\alias{addSubs}
\title{Add substances to the environment}
\usage{
addSubs(object, smax = 0, mediac = object@mediac)
}
\arguments{
\item{object}{An object of class Arena.}

\item{smax}{A number indicating the maximum substance concentration per grid cell.}

\item{mediac}{A character vector giving the names of substances, which should be added to the environment (the default takes all possible substances).}
}
\description{
The generic function \code{addSubs} adds specific substances to the environment.
}
\details{
If nothing but \code{object} is given, then all possible substrates are initilized with a concentration of 0. Afterwards, \code{\link{changeSub} can be used to modify the concentrations of specific substances.}
}
\examples{
\dontrun{
ecore <- model #get Escherichia coli core metabolic model
bac <- Bac(ecore,deathrate=0.05,duplirate=0.5,
           growthlimit=0.05,growtype="exponential") #initialize a bacterium
arena <- Arena(20,20) #initialize the environment
addSubs(arena,20,c("EX_glc(e)","EX_o2(e)","EX_pi(e)")) #add substances glucose, oxygen and phosphate
}
}
\seealso{
\code{\link{Arena-class}} and \code{\link{changeSub}}
}

