% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Arena.R
\docType{class}
\name{Arena-class}
\alias{Arena-class}
\title{Structure of the S4 class "Arena"}
\description{
Structure of the S4 class \code{Arena} to represent the environment in which Organisms and Substances interact.
}
\section{Slots}{

\describe{
\item{\code{orgdat}}{A data frame collecting information about the accumulated growth, type, phenotype, x and y position for each individual in the environment.}

\item{\code{specs}}{A list of organism types and their associated parameters.}

\item{\code{media}}{A list of objects of class \code{\link{Substance-class}} for each compound in the environment.}

\item{\code{phenotypes}}{A list of unique phenotypes (metabolites consumed and produced), which occurred in the environment.}

\item{\code{mediac}}{A character vector containing the names of all substances in the environment.}

\item{\code{occmat}}{A sparse matrix showing which cells in the environment are occupied by individuals.}

\item{\code{tstep}}{A number giving the time (in h) per iteration.}

\item{\code{stir}}{A boolean variable indicating if environment should be stirred.}

\item{\code{n}}{A number giving the horizontal size of the environment.}

\item{\code{m}}{A number giving the vertical size of the environment.}
}}

