% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\name{Arena-constructor}
\alias{Arena-constructor}
\alias{Arena}
\title{Constructor of the S4 class \code{\link{Arena-class}}}
\usage{
Arena(Lx = NULL, Ly = NULL, n = 100, m = 100, seed = sample(1:10000, 1), ...)
}
\arguments{
\item{Lx}{A number giving the horizontal grid size in cm.}

\item{Ly}{A number giving the vertical grid size in cm.}

\item{n}{A number giving the horizontal size of the environment.}

\item{m}{A number giving the vertical size of the environment.}

\item{seed}{An integer refering to the random number seed used to be reproducible}

\item{...}{Arguments of \code{\link{Arena-class}}}
}
\description{
Constructor of the S4 class \code{\link{Arena-class}}
}
