% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organism.R
\docType{class}
\name{Bac-class}
\alias{Bac-class}
\title{Structure of the S4 class "Bac"}
\description{
Structure of the S4 class \code{Bac} inheriting from class \code{\link{Organism-class}} representing bacterial cells.
}
\section{Slots}{

\describe{
\item{\code{chem}}{A character vector indicating name of substance which is the chemotaxis attractant. Empty character vector if no chemotaxis.}
}}

