% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BacArena.R
\docType{package}
\name{BacArena}
\alias{BacArena-package}
\alias{BacArena}
\title{BacArena: An Agent-Based Modeling Framework for Cellular Communities}
\description{
The BacArena package provides six classes: Arena (subclass Eval), Organism (subclasses Bac, Human) and Substance.
Accordingly there are three categories of important functions:
Arena, Organism and Substance.
}
\section{Arena functions}{

The Arena functions ...
}

\section{Organism functions}{

The Organism functions ...
}

\section{Substance functions}{

The Substance functions ...
}

\seealso{
Useful links:
\itemize{
  \item \url{https://BacArena.github.io/}
  \item Report bugs at \url{https://github.com/euba/BacArena/issues}
}

}
\author{
\strong{Maintainer}: Johannes Zimmermann \email{j.zimmermann@iem.uni-kiel.de}

Authors:
\itemize{
  \item Eugen Bauer \email{eugen.bauer@uni.lu}
}

}
