% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Organism.R
\name{budging}
\alias{budging}
\title{Function for budging of fellow bacteria, while one is moving}
\usage{
budging(object, population, j, hood, repli = F)
}
\arguments{
\item{object}{An object of class Bac.}

\item{population}{An object of class Arena.}

\item{j}{The number of the iteration of interest.}

\item{hood}{A vector giving the Moore neighbourhood of the bacterium of interest}
}
\value{
Returns the updated enivironment of the \code{population} parameter with all new positions of individuals on the grid.
}
\description{
The generic function \code{budging} implements a random bacterial movement in the Moore neighbourhood while pushing other bacteria into the next grid cell
}
\details{
Bacteria move to a random position in the Moore neighbourhood. If the position is already occupied by another individual, the individual gets push away to the next grid position, which pushes another individual until a free position is reached. If the borders of the environment are reached, then the first individual is not able to move and every subsequently pushed bacterium stays at its original position.
}
\examples{
NULL
}
\seealso{
\code{\link{Bac-class}}, \code{\link{getHood}}, \code{\link{emptyHood}} and \code{\link{move}}
}

