% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Arena.R
\name{changeOrg}
\alias{changeOrg}
\title{Change organisms in the environment}
\usage{
changeOrg(object, neworgdat)
}
\arguments{
\item{object}{An object of class Arena.}

\item{neworgdat}{A data frame with new information about the accumulated growth, type, phenotype, x and y position for each individual in the environment.}
}
\description{
The generic function \code{changeOrg} changes organisms in the environment.
}
\details{
The argument \code{neworgdat} contains the same information as the \code{orgdat} slot of \code{\link{Arena-class}}. The \code{orgdat} slot of an \code{Arena} object can be used to create \code{neworgdat}.
}
\examples{
\dontrun{
ecore <- model #get Escherichia coli core metabolic model
bac <- Bac(ecore,deathrate=0.05,duplirate=0.5,
           growthlimit=0.05,growtype="exponential") #initialize a bacterium
arena <- Arena(20,20) #initialize the environment
addOrg(arena,bac,amount=10) #add 10 organisms
neworgdat <- orgdat(arena) #get the current orgdat
neworgdat <- neworgdat[-1,] #remove the first individual
changeOrg(arena,neworgdat)
}
}
\seealso{
\code{\link{Arena-class}} and \code{\link{addOrg}}
}

