% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\name{checkPhen_par}
\alias{checkPhen_par}
\alias{checkPhen_par,Arena-method}
\title{Function for checking phenotypes in the environment}
\usage{
checkPhen_par(object, org, cutoff = 1e-06, fbasol)

\S4method{checkPhen_par}{Arena}(object, org, cutoff = 1e-06, fbasol)
}
\arguments{
\item{object}{An object of class Arena.}

\item{org}{An object of class Organism.}

\item{cutoff}{A number giving the cutoff for values of the objective function and fluxes of exchange reactions.}

\item{fbasol}{Problem object according to the constraints and then solved with \code{optimizeProb}.}
}
\description{
The generic function \code{checkPhen_par} checks and adds the phenotypes of organisms in the environment.
}
