% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Organism.R
\name{chemotaxis}
\alias{chemotaxis}
\title{Function for chemotaxis of bacteria to their prefered substrate}
\usage{
chemotaxis(object, population, j)
}
\arguments{
\item{object}{An object of class Bac.}

\item{population}{An object of class Arena.}

\item{j}{The number of the iteration of interest.}
}
\description{
The generic function \code{chemotaxis} implements a bacterial movement in the Moore neighbourhood to the highest substrate concentration.
}
\details{
Bacteria move to a position in the Moore neighbourhood which has the highest concentration of the prefered substrate, which is not occupied by other individuals. The prefered substance is given by slot \code{chem} in the \code{Bac} object. If there is no free space the individuals stays in the same position. If the concentration in the Moore neighbourhood has the same concentration in every position, then random movement is implemented.
}
\examples{
\dontrun{
ecore <- model #get Escherichia coli core metabolic model
bac <- Bac(ecore,deathrate=0.05,duplirate=0.5,
           growthlimit=0.05,growtype="exponential") #initialize a bacterium
arena <- Arena(20,20) #initialize the environment
addOrg(arena,bac,amount=10) #add 10 organisms
addSubs(arena,40) #add all possible substances
chemotaxis(bac,arena,1)
}
}
\seealso{
\code{\link{Bac-class}}, \code{\link{getHood}} and \code{\link{emptyHood}}
}

