% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colpal1}
\alias{colpal1}
\title{Color palette}
\format{A vector with 269 color hex-codes}
\source{
\url{http://godsnotwheregodsnot.blogspot.ru/2013/11/kmeans-color-quantization-seeding.html}
}
\usage{
colpal1
}
\description{
color dictionary of 269 maximally distinct colors from all previous colors
}
\keyword{datasets}
