% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colpal2}
\alias{colpal2}
\title{Color palette}
\format{A vector with 20 color hex-codes}
\source{
\url{http://graphicdesign.stackexchange.com/questions/3682/where-can-i-find-a-large-palette-set-of-contrasting-colors-for-coloring-many-d}
}
\usage{
colpal2
}
\description{
20 optimally distinct colors
}
\keyword{datasets}
