% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colpal3}
\alias{colpal3}
\title{Color palette}
\format{A vector with 22 color hex-codes}
\source{
\url{http://www.iscc-archive.org/pdf/PC54_1724_001.pdf}
}
\usage{
colpal3
}
\description{
K. Kelly (1965): Twenty-two colors of maximum contrast. // Color Eng., 3(6), 1965
}
\keyword{datasets}
