% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colpal4}
\alias{colpal4}
\title{Color palette}
\format{A vector with 26 color hex-codes}
\source{
\url{http://eeecon.uibk.ac.at/~zeileis/papers/Zeileis+Hornik+Murrell-2009.pdf}
}
\usage{
colpal4
}
\description{
26 distinct colors (Zeileis(2009): Escaping RGBland: Selecting Colors for Statistical Graphics)
}
\keyword{datasets}
