% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Organism.R
\name{constrain}
\alias{constrain}
\title{Function for constraining the models based on metabolite concentrations}
\usage{
constrain(object, reacts, lb, dryweight, time)
}
\arguments{
\item{object}{An object of class Organisms.}

\item{reacts}{A character vector giving the names of reactions which should be constrained.}

\item{lb}{A numeric vector giving the constraint values of lower bounds.}

\item{dryweight}{A number giving the current dryweight of the organism.}

\item{time}{A number giving the time intervals for each simulation step.}
}
\value{
Returns the lower bounds, which carry the constraints and names of relevant reactions.
}
\description{
The generic function \code{constrain} changes the constraints of the model representation of an organism.
}
\details{
The constraints are calculated according to the flux definition as mmol/(gDW*hr) with the parameters \code{dryweight} and \code{time}.
}
\examples{
\dontrun{
ecore <- model #get Escherichia coli core metabolic model
org <- Organism(ecore,deathrate=0.05,duplirate=0.5,
           growthlimit=0.05,growtype="exponential") #initialize a organism
lobnds <- constrain(org,medium(org),lbnd(org)[medium(org)],1,1)
}
}
\seealso{
\code{\link{Organism-class}}
}

