% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Organism.R
\name{consume}
\alias{consume}
\title{Function to account for the consumption and production of substances}
\usage{
consume(object, sublb, cutoff = 1e-06)
}
\arguments{
\item{object}{An object of class Organisms.}

\item{sublb}{A vector containing the substance concentrations in the current position of the individual of interest.}

\item{cutoff}{A number giving the cutoff value by which value of objective function is considered greater than 0.}
}
\value{
Returns the updated vector containing the substance concentrations in the current position of the individual of interest.
}
\description{
The generic function \code{consume} implements the consumption and production of substances based on the flux of exchange reactions of organisms
}
\details{
The consumption is implemented by adding the flux of the exchange reactions to the current substance concentrations.
}
\examples{
NULL
}
\seealso{
\code{\link{Organism-class}}
}

