% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Arena.R
\name{createGradient}
\alias{createGradient}
\title{Change substance concentration patterns in the environment according to a gradient}
\usage{
createGradient(object, mediac, position, smax, steep)
}
\arguments{
\item{object}{An object of class Arena.}

\item{mediac}{A character vector giving the names of substances, which should be added to the environment (the default takes all possible substances).}

\item{position}{A character vector giving the position (top, bottom, right and left) of the gradient.}

\item{smax}{A number giving the maximum concentration of the substance.}

\item{steep}{A number between 0 and 1 giving the steepness of the gradient (concentration relative to the arena size).}
}
\description{
The generic function \code{createGradient} changes specific substance concentration patterns in the environment.
}
\details{
This function can be used to add gradients of specific substances in the environment.
}
\examples{
\dontrun{
ecore <- model #get Escherichia coli core metabolic model
bac <- Bac(ecore,deathrate=0.05,duplirate=0.5,
           growthlimit=0.05,growtype="exponential") #initialize a bacterium
arena <- Arena(20,20) #initialize the environment
addOrg(arena,bac,amount=10) #add 10 organisms
addSubs(arena,30) #add all substances with no concentrations.
createGradient(arena,smax=50,mediac=c("EX_glc(e)","EX_o2(e)","EX_pi(e)"),
             position='top',steep=0.5)
}
}
\seealso{
\code{\link{Arena-class}} and \code{\link{changeSubs}}
}

