% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Arena.R
\name{dat2mat}
\alias{dat2mat}
\title{Function for transforming the organism data frame to a presence/absence matrix of organisms}
\usage{
dat2mat(object)
}
\arguments{
\item{object}{An object of class Arena.}
}
\value{
Returns the presence/absence matrix of organisms on the grid based on the \code{orgdat} slot of the \code{Arena} class.
}
\description{
The generic function \code{dat2mat} simulates the event of mixing all substrates and organisms in the environment.
}
\examples{
\dontrun{
ecore <- model #get Escherichia coli core metabolic model
bac <- Bac(ecore,deathrate=0.05,duplirate=0.5,
           growthlimit=0.05,growtype="exponential") #initialize a bacterium
arena <- Arena(20,20) #initialize the environment
addOrg(arena,bac,amount=10) #add 10 organisms
occmat <- dat2mat(arena)
image(occmat)
}
}
\seealso{
\code{\link{Arena-class}} and \code{\link{getSublb}}
}

