% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\name{diffuse_par}
\alias{diffuse_par}
\alias{diffuse_par,Arena-method}
\title{Function for parallelzied diffusion}
\usage{
diffuse_par(object, lrw, cluster_size, sublb)

\S4method{diffuse_par}{Arena}(object, lrw, cluster_size, sublb)
}
\arguments{
\item{object}{An object of class Arena.}

\item{lrw}{A numeric value needed by solver to estimate array size (by default lwr is estimated in the simEnv() by the function estimate_lrw())}

\item{cluster_size}{Amount of cores to be used}

\item{sublb}{A matrix with the substrate concentration for every individual in the environment based on their x and y position.}
}
\description{
The generic function \code{diffuse_par} computes the media distribution via diffusion in parallel
}
