% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Substance.R
\name{diffuseCpp}
\alias{diffuseCpp}
\title{Function for naive diffusion (neighbourhood) of the Substance matrix}
\usage{
diffuseCpp(object)
}
\arguments{
\item{object}{An object of class Substance.}
}
\description{
The generic function \code{diffuseR} implements the diffusion in the Moore neighbourhood with \code{C++} code.
}
\details{
The function is a wrapper of the \code{C++} function \code{diffuseNaiveCpp}. The diffusion is implemented by iterating through each cell in the grid and taking the cell with the lowest concentration in the Moore neighbourhood to update the concentration of both by their mean.
}
\examples{
\dontrun{
sub <- Substance(n=20,m=20,smax=40,name='test',difunc='cpp') #initialize test substance
diffuseCpp(sub)
}
}
\seealso{
\code{\link{Substance-class}} and \code{\link{diffuseR}}
}

