% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Substance.R
\name{diffusePDE}
\alias{diffusePDE}
\title{Function for diffusion using a partial differential equation}
\usage{
diffusePDE(object)
}
\arguments{
\item{object}{An object of class Substance.}
}
\description{
The generic function \code{diffusePDE} implements a partial differential equation after Grajdeanu (2007).
}
\details{
The function is a wrapper of the \code{C++} function \code{diffuseGrajdeanuCpp}.
}
\examples{
\dontrun{
sub <- Substance(n=20,m=20,smax=40,name='test',difunc='pde') #initialize test substance
diffusePDE(sub)
}
}
\seealso{
\code{\link{Substance-class}} and \code{\link{diffuseR}}
}

