% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Organism.R
\name{emptyHood}
\alias{emptyHood}
\title{Function to check if the there is a free place in the Moore neighbourhood}
\usage{
emptyHood(object, occmat, x, y)
}
\arguments{
\item{object}{An object of class Organisms.}

\item{occmat}{A matrix giving where the individuals in environment are occupying a specific position.}

\item{x}{A number giving the x position of the individual of interest in its environment.}

\item{y}{A number giving the y position of the individual of interest in its environment.}
}
\value{
Returns the free position in the Moore neighbourhood, which is not occupied by other individuals. If there is no free space \code{NULL} is returned.
}
\description{
The generic function \code{emptyHood} gives a free space which is present in the Moore neighbourhood of an individual of interest.
}
\examples{
NULL
}
\seealso{
\code{\link{Organism-class}} and \code{\link{getHood}}
}

