% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Arena.R
\docType{class}
\name{Eval-class}
\alias{Eval-class}
\title{Structure of the S4 class "Eval"}
\description{
Structure of the S4 class \code{Eval} inheriting from class \code{\link{Arena}} for the analysis of simulations.
}
\section{Slots}{

\describe{
\item{\code{medlist}}{A list of compressed medium concentrations (only changes of concentrations are stored) per time step.}

\item{\code{simlist}}{A list of the organism features per time step.}

\item{\code{subchange}}{A vector of all substrates with numbers indicating the degree of change in the overall simulation.}
}}

