% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\name{Eval-constructor}
\alias{Eval-constructor}
\alias{Eval}
\title{Constructor of the S4 class \code{\link{Eval-class}}}
\usage{
Eval(arena)
}
\arguments{
\item{arena}{An object of class Arena.}
}
\description{
Constructor of the S4 class \code{\link{Eval-class}}
}
