% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Arena.R
\name{evalArena}
\alias{evalArena}
\title{Function for plotting spatial and temporal change of populations and/or concentrations}
\usage{
evalArena(object, plot_items = "population", phencol = F, retdata = F)
}
\arguments{
\item{object}{An object of class Eval.}

\item{plot_items}{A character vector giving the items, which should be plotted.}

\item{phencol}{A boolean variable indicating if the phenotypes of the organisms in the environment should be integrated as different colors in the population plot.}

\item{retdata}{A boolean variable indicating if the data used to generate the plots should be returned.}
}
\value{
Returns several plots of the chosen plot items. Optional the data to generate the original plots can be returned.
}
\description{
The generic function \code{evalArena} plots heatmaps from the simulation steps in an \code{Eval} object.
}
\details{
If \code{phencol} is \code{TRUE} then different phenotypes of the same organism are visualized by varying colors, otherwise different organism types are represented by varying colors. The parameter \code{retdata} can be used to access the data used for the returned plots to create own custom plots.
}
\examples{
\dontrun{
ecore <- model #get Escherichia coli core metabolic model
bac <- Bac(ecore,deathrate=0.05,duplirate=0.5,
           growthlimit=0.05,growtype="exponential") #initialize a bacterium
arena <- Arena(20,20) #initialize the environment
addOrg(arena,bac,amount=10) #add 10 organisms
addSubs(arena,40) #add all possible substances
eval <- simEnv(arena,10)
evalArena(eval)
## if animation package is installed a movie of the simulation can be stored:
library(animation)
saveVideo({evalArena(eval)},video.name="Ecoli_sim.mp4")
}
}
\seealso{
\code{\link{Eval-class}} and \code{\link{Arena-class}}
}

