% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\docType{methods}
\name{findFeeding}
\alias{findFeeding}
\alias{findFeeding,Eval-method}
\title{Function for investigation of feeding between phenotypes}
\usage{
findFeeding(object, dict = NULL, tcut = 5, scut = list(),
  legendpos = "topleft")

\S4method{findFeeding}{Eval}(object, dict = NULL, tcut = 5, scut = list(),
  legendpos = "topleft")
}
\arguments{
\item{object}{An object of class Eval.}

\item{dict}{List defining new substance names. List entries are intepreted as old names and the list names as the new ones.}

\item{tcut}{Integer giving the minimal mutual occurence ot be considered (dismiss very seldom feedings)}

\item{scut}{List of substance names which should be ignored}

\item{legendpos}{A character variable declaring the position of the legend}
}
\value{
Graph (igraph)
}
\description{
The generic function \code{findFeeding}
}

