% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\name{findFeeding2}
\alias{findFeeding2}
\alias{findFeeding2,Eval-method}
\title{Function for investigation of feeding between phenotypes}
\usage{
findFeeding2(object, time, mets, rm_own = T, ind_threshold = 0, collapse = F)

\S4method{findFeeding2}{Eval}(object, time, mets, rm_own = T, ind_threshold = 0, collapse = F)
}
\arguments{
\item{object}{An object of class Eval.}

\item{time}{A numeric vector giving the simulation steps which should be plotted.}

\item{mets}{Character vector of substance names which should be considered}

\item{rm_own}{A boolean flag indicating if interactions within same species should be plotted}

\item{ind_threshold}{A number indicating the threshold of individuals to be considered as producers/consumers}

\item{collapse}{A boolean flag indicating if all phenotypes for every species should be collapsed to either producers or consumers}
}
\value{
Graph (igraph)
}
\description{
The generic function \code{findFeeding2}
}
