% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\docType{methods}
\name{findFeeding3}
\alias{findFeeding3}
\alias{findFeeding3,Eval-method}
\title{Function for investigation of feeding between phenotypes}
\usage{
findFeeding3(object, time, mets)

\S4method{findFeeding3}{Eval}(object, time, mets)
}
\arguments{
\item{object}{An object of class Eval.}

\item{time}{A numeric vector giving the simulation steps which should be plotted.}

\item{mets}{Character vector of substance names which should be considered}
}
\value{
Graph (igraph)
}
\description{
The generic function \code{findFeeding3}
}

