% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organism.R
\name{findrBiomass}
\alias{findrBiomass}
\title{Find biomass reaction in model}
\usage{
findrBiomass(model, keys = c("biom", "cpd11416"))
}
\arguments{
\item{model}{Object of class sybil::modelorg containging the genome sclae metabolic model}

\item{keys}{Vector with strings which are used to find biomass reaction in model}
}
\value{
Vector with reaction ids for biomass reaction(s)
}
\description{
Helper function to search for biomass reaction in available reactions of a model
}
