% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Organism.R
\name{getHood}
\alias{getHood}
\title{Function to get Moore-neighbourhood of a organism together with its relative position}
\usage{
getHood(object, occmat, x, y)
}
\arguments{
\item{object}{An object of class Organisms.}

\item{occmat}{A matrix giving where the individuals in environment are occupying a specific position.}

\item{x}{A number giving the x position of the individual of interest in its environment.}

\item{y}{A number giving the y position of the individual of interest in its environment.}
}
\value{
Returns a matrix giving the Moore neighbourhood with positions that are occupied by other individuals.
}
\description{
The generic function \code{getHood} gives the Moore neighbourhood of an individual of interest.
}
\examples{
NULL
}
\seealso{
\code{\link{Organism-class}} and \code{\link{emptyHood}}
}

