% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\docType{methods}
\name{getSubHist}
\alias{getSubHist}
\alias{getSubHist,Eval-method}
\title{Function to get timeline of a substance}
\usage{
getSubHist(object, sub)

\S4method{getSubHist}{Eval}(object, sub)
}
\arguments{
\item{object}{An object of class Eval.}

\item{sub}{Name of a substance.}
}
\description{
The generic function \code{getSubHist} returns list with amount of substance for each timestep
}
