% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Arena.R
\name{getSublb}
\alias{getSublb}
\title{Function for calculated the substrate concentration for every organism}
\usage{
getSublb(object)
}
\arguments{
\item{object}{An object of class Arena.}
}
\value{
Returns the substrate concentration for every individual in the environment with substrates as well as x and y positions as columns and rows for each organism.
}
\description{
The generic function \code{getSublb} calculates the substrate concentration for every individual in the environment based on their x and y position.
}
\examples{
\dontrun{
ecore <- model #get Escherichia coli core metabolic model
bac <- Bac(ecore,deathrate=0.05,duplirate=0.5,
           growthlimit=0.05,growtype="exponential") #initialize a bacterium
arena <- Arena(20,20) #initialize the environment
addOrg(arena,bac,amount=10) #add 10 organisms
addSubs(arena,40) #add all possible substances
sublb <- getSublb(arena)
}
}
\seealso{
\code{\link{Arena-class}}
}

