% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\docType{methods}
\name{getVarSubs}
\alias{getVarSubs}
\alias{getVarSubs,Eval-method}
\title{Function to get varying substances}
\usage{
getVarSubs(object, only_products = TRUE, only_substrates = TRUE,
  cutoff = 1e-06)

\S4method{getVarSubs}{Eval}(object, only_products = FALSE,
  only_substrates = FALSE, cutoff = 1e-06)
}
\arguments{
\item{object}{An object of class Eval.}

\item{only_products}{A boolean indicating if only products should be shown}

\item{only_substrates}{A boolean indicating if only substrates should be shown}

\item{cutoff}{Value used to define numeric accuracy while interpreting optimization results}
}
\description{
The generic function \code{getVarSubs} returns ordered list of substances that showed variance during simulation
}

