% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organism.R
\name{Human-constructor}
\alias{Human-constructor}
\alias{Human}
\title{Constructor of the S4 class \code{\link{Human-class}}}
\usage{
Human(
  model,
  objective = model@react_id[which(model@obj_coef != 0)],
  speed = 0,
  ...
)
}
\arguments{
\item{model}{model}

\item{objective}{A character vector representing the current reaction which should be used as an objective function for the flux balance analysis.}

\item{speed}{A integer vector representing the speed by which bacterium is moving (given by cell per iteration).}

\item{...}{Arguments of \code{\link{Organism}}}
}
\value{
Object of class \code{\link{Human-class}}
}
\description{
Constructor of the S4 class \code{\link{Human-class}}
}
