% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organism.R
\docType{class}
\name{Human-class}
\alias{Human-class}
\title{Structure of the S4 class "Human"}
\description{
Structure of the S4 class \code{Human} inheriting from class \code{\link{Organism-class}} representing human cells.
}
\section{Slots}{

\describe{
\item{\code{objective}}{A character vector representing the current reaction which should be used as an objective function for the flux balance analysis.}
}}

