% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Arena.R
\name{minePheno}
\alias{minePheno}
\title{Function for mining/analyzing phenotypes which occured on the arena}
\usage{
minePheno(object, plot_type = "pca")
}
\arguments{
\item{object}{An object of class Eval.}

\item{plot_type}{A character vector giving the plot which should be returned (either "pca" for a principle coordinate analysis or "hclust" for hierarchical clustering).}
}
\value{
Returns a plot for each simulation step representing the similarity of phenotypes of organisms within the environment.
}
\description{
The generic function \code{minePheno} mines the similarity and differences of phenotypes reconstructed by \code{getPhenoMat} for each simulation step in an \code{Eval} object.
}
\details{
The phenotypes are defined by flux through exchange reactions, which indicate potential differential substrate usages.
}
\examples{
\dontrun{
ecore <- model #get Escherichia coli core metabolic model
bac <- Bac(ecore,deathrate=0.05,duplirate=0.5,
           growthlimit=0.05,growtype="exponential") #initialize a bacterium
arena <- Arena(20,20) #initialize the environment
addOrg(arena,bac,amount=10) #add 10 organisms
addSubs(arena,40) #add all possible substances
eval <- simulate(arena,10)
minePheno(eval)
}
}
\seealso{
\code{\link{Eval-class}} and \code{\link{getPhenoMat}}
}

