% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Organism.R
\name{move}
\alias{move}
\title{Function for random movement of bacteria}
\usage{
move(object, population, j)
}
\arguments{
\item{object}{An object of class Bac.}

\item{population}{An object of class Arena.}

\item{j}{The number of the iteration of interest.}
}
\description{
The generic function \code{move} implements a random movement in the Moore neighbourhood of an individual.
}
\details{
Bacteria move in a random position the Moore neighbourhood, which is not occupied by other individuals. If there is no free space the individuals stays in the same position.
}
\examples{
\dontrun{
ecore <- model #get Escherichia coli core metabolic model
bac <- Bac(ecore,deathrate=0.05,duplirate=0.5,
           growthlimit=0.05,growtype="exponential") #initialize a bacterium
arena <- Arena(20,20) #initialize the environment
addOrg(arena,bac,amount=10) #add 10 organisms
addSubs(arena,40) #add all possible substances
move(bac,arena,1)
}
}
\seealso{
\code{\link{Bac-class}}, \code{\link{getHood}} and \code{\link{emptyHood}}
}

