% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organism.R
\name{NemptyHood}
\alias{NemptyHood}
\alias{NemptyHood,Organism-method}
\title{Function to check if the there is a free place in the Moore neighbourhood}
\usage{
NemptyHood(object, pos, n, m, x, y, occupyM, inverse = FALSE)

\S4method{NemptyHood}{Organism}(object, pos, n, m, x, y, occupyM, inverse = FALSE)
}
\arguments{
\item{object}{An object of class Organisms.}

\item{pos}{A dataframe with all occupied x and y positions}

\item{n}{A number giving the horizontal size of the environment.}

\item{m}{A number giving the vertical size of the environment.}

\item{x}{A number giving the x position of the individual of interest in its environment.}

\item{y}{A number giving the y position of the individual of interest in its environment.}

\item{occupyM}{A matrix indicating grid cells that are obstacles}

\item{inverse}{Return occupied positions instead}
}
\value{
Returns the free position in the Moore neighbourhood, which is not occupied by other individuals. If there is no free space \code{NULL} is returned.
}
\description{
The generic function \code{NemptyHood} gives a free space which is present in the Moore neighbourhood of an individual of interest.
}
\examples{
NULL
}
\seealso{
\code{\link{Organism-class}}
}
