% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{openArena}
\alias{openArena}
\title{Start simulation}
\usage{
openArena()
}
\value{
Returns an object of class \code{Eval} which can be used for subsequent analysis steps.
}
\description{
The function \code{openArena} can be used to start a default simulation.
}
\examples{
\donttest{ 
sim <- openArena()
evalArena(sim, time=5, phencol = TRUE, 
          plot_items=c("Population", "EX_o2(e)", "EX_for(e)", "EX_glc(e)"))
}
}
