% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organism.R
\name{Organism-constructor}
\alias{Organism}
\alias{Organism-constructor}
\title{Constructor of the S4 class \code{Organism}}
\usage{
Organism(model, algo = "fba", ex = "EX_", ex_comp = NA,
  csuffix = "\\\\[c\\\\]", esuffix = "\\\\[e\\\\]", lyse = F,
  feat = list(), typename = NA, ...)
}
\arguments{
\item{model}{model}

\item{algo}{A single character string giving the name of the algorithm to use. See \link[sybil]{SYBIL_SETTINGS}}

\item{ex}{Identifier for exchange reactions}

\item{ex_comp}{ex_comp}

\item{csuffix}{csuffix}

\item{esuffix}{esuffix}

\item{lyse}{A boolean variable indicating if the organism should lyse after death.}

\item{feat}{A list containing conditional features for the object (contains at the momement only biomass components for lysis).}

\item{typename}{A string defining the name (set to model name in default case)}

\item{...}{Arguments of \code{\link{Organism-class}}}
}
\value{
Object of class Organism
}
\description{
The constructor to get a new object of class \code{Organism}
}

