% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{plotAbundance}
\alias{plotAbundance}
\alias{plotSubVar}
\title{Plot abundances of species}
\usage{
plotAbundance(simlist, time = c(NULL, NULL), col = colpal3,
  return_dat = F)

plotSubVar(simlist, metsel)
}
\arguments{
\item{simlist}{A list of simulations (eval objects).}

\item{time}{A vector with start and end time to be considered (default: total time)}

\item{col}{Vector with color that should be used}

\item{return_dat}{Should plain text mean abundances be returned? (default false)}

\item{simlist}{A list of simulations (eval objects).}

\item{metlist}{A vector with the name of exchange reactions of interest}
}
\description{
The function \code{plotAbundance} takes a list of simulations and return a boxplot with species abundances

The function \code{plotSubVar} takes a list of simulations and return a barplot with most varying substances
}

