% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{plotFluxVar}
\alias{plotFluxVar}
\title{Plot population flux variations}
\usage{
plotFluxVar(simlist, metsel)
}
\arguments{
\item{simlist}{A list of simulations (eval objects).}

\item{metsel}{A vector with the name of exchange reactions of interest}
}
\description{
The function \code{plotFluxVar} takes a list of simulations and metabolites, returning a plot with metabolite fluxes for each species
}
