% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{plotFVA}
\alias{plotFVA}
\title{Function to plot population level minimum and maximum flux from alternative optimal solutions obtained by FVA}
\usage{
plotFVA(fvares, mediac)
}
\arguments{
\item{fvares}{results of FVA results to plot, obtained from function fluxVarSim}

\item{mediac}{List with substances.}
}
\description{
The generic function \code{plotFVA} plots population level flux results obtained from function fluxVarSim
}
\details{
Returns ggplot objects
}
