% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{plotGrowthCurve}
\alias{plotGrowthCurve}
\title{Plot growth curve for several simulations}
\usage{
plotGrowthCurve(simlist)
}
\arguments{
\item{simlist}{A list of simulations (eval objects).}
}
\description{
The function \code{plotGrowthCurve} takes a list of simulations and plots the time course of species with standard deviation.
}

