% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{plotInterNum}
\alias{plotInterNum}
\title{Plot number of variation in number of interactions for several simulations}
\usage{
plotInterNum(simlist, title = "Variation in number of interactions", size = 1)
}
\arguments{
\item{simlist}{A list of simulations (eval objects).}

\item{title}{Title of the plot}

\item{size}{A scaling factor for plot text and line size}
}
\description{
The function \code{plotInterNum} takes a list of simulations and plots the time course of the number of metabolic interactions with standard deviation.
}
