% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{plotPhenCurve}
\alias{plotPhenCurve}
\title{Plot growth curve for several simulations}
\usage{
plotPhenCurve(simlist, subs, phens = NULL, time = c(NULL, NULL),
  ret_phengroups = FALSE, cluster = TRUE, col = colpal3)
}
\arguments{
\item{simlist}{A list of simulations (eval objects).}

\item{subs}{A vector of substance names that are used for phenotype clustering.}

\item{phens}{If phencurve is given then phens specifies the phenotypes which sould be plotted again.}

\item{time}{Vector with two entries defining start and end time}

\item{ret_phengroups}{True if clustered phenotype groups should be returned.}

\item{cluster}{True phenotypes should be clustered/condensed.}

\item{col}{Vector with color that should be used}
}
\description{
The function \code{plotPhenCurve} takes a list of simulations and plots the time course of species with standard deviation.
}

