% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{plotPhenNum}
\alias{plotPhenNum}
\title{Plot number of phenotypes curve for several simulations}
\usage{
plotPhenNum(simlist, title = "Phenotype number variation", size = 1)
}
\arguments{
\item{simlist}{A list of simulations (eval objects).}

\item{title}{Title of the plot}

\item{size}{A scaling factor for plot text and line size}
}
\description{
The function \code{plotPhenNum} takes a list of simulations and plots the time course of the number of phenotypes with standard deviation.
}
