% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{plotReaActivity}
\alias{plotReaActivity}
\title{Function to plot reaction activity for every species}
\usage{
plotReaActivity(
  simlist,
  reactions = list(),
  spec_list = NULL,
  ret_data = FALSE
)
}
\arguments{
\item{simlist}{An object of class Eval or a list with objects of class Eval.}

\item{reactions}{List of reaction names}

\item{spec_list}{List of species names to be considered (default all)}

\item{ret_data}{Set true if data should be returned}
}
\description{
The generic function \code{plotReaActivity} displays the usage of reactions for all species
}
\details{
Returns ggplot objects
}
