% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\name{plotShadowCost}
\alias{plotShadowCost}
\alias{plotShadowCost,Eval-method}
\title{Function to plot substance shadow costs for a specie}
\usage{
plotShadowCost(
  object,
  spec_nr = 1,
  sub_nr = 10,
  cutoff = -1,
  noplot = FALSE,
  useNames = FALSE
)

\S4method{plotShadowCost}{Eval}(
  object,
  spec_nr = 1,
  sub_nr = 10,
  cutoff = -1,
  noplot = FALSE,
  useNames = FALSE
)
}
\arguments{
\item{object}{An object of class Eval.}

\item{spec_nr}{Number of the specie}

\item{sub_nr}{Maximal number of substances to be show}

\item{cutoff}{Shadow costs should be smaller than cutoff}

\item{noplot}{Do not plot}

\item{useNames}{Use substance names instead of ids}
}
\description{
The generic function \code{plotShadowCost} plots substances have the highest impact on further growth (shadow cost < 0)
}
\details{
Returns ggplot objects
}
