% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\docType{methods}
\name{plotSpecActivity}
\alias{plotSpecActivity}
\alias{plotSpecActivity,Eval-method}
\title{Function to plot substance usage for a single species}
\usage{
plotSpecActivity(object, spec_nr = 1, var_nr = 10)

\S4method{plotSpecActivity}{Eval}(object, spec_nr = 1, var_nr = 10)
}
\arguments{
\item{object}{An object of class Eval.}

\item{spec_nr}{Number of the species}

\item{var_nr}{Number of most varying substances to be used}
}
\description{
The generic function \code{plotSpecActivity} displays the input/output substances with the highest variance for certain species
}
\details{
Returns ggplot objects
}

