% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{plotSubCurve}
\alias{plotSubCurve}
\title{Plot substance curve for several simulations}
\usage{
plotSubCurve(simlist, mediac = NULL)
}
\arguments{
\item{simlist}{A list of simulations (eval objects).}

\item{mediac}{A vector of substances (if not specified most varying substances will be taken.)}
}
\description{
The function \code{plotSubCurve} takes a list of simulations and plots the time course of substances with standard deviation.
}

