% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\name{plotSubDist2}
\alias{plotSubDist2}
\alias{plotSubDist2,Eval-method}
\title{Function to overview the spatial distribution of a substance over time.}
\usage{
plotSubDist2(object, sub, times = NULL)

\S4method{plotSubDist2}{Eval}(object, sub, times = NULL)
}
\arguments{
\item{object}{An object of class Eval.}

\item{sub}{Name of a substance.}

\item{times}{Time points to be considered.}
}
\description{
The generic function \code{plotSubDist2} returns a plot for every time step which shows the substance concentration in the environment.
}
\details{
Returns a plot with
}
