% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{plotSubUsage}
\alias{plotSubUsage}
\title{Function to plot usage of substances species wise}
\usage{
plotSubUsage(simlist, subs = vector(), cutoff = 0.01, ret_data = FALSE)
}
\arguments{
\item{simlist}{An object of class Eval or a list with objects of class Eval.}

\item{subs}{List of substance names}

\item{cutoff}{Total values below cutoff will be dismissed}

\item{ret_data}{Set true if data should be returned}
}
\description{
The generic function \code{plotSubUsage} displays for given substances the quantities of absorption and production for each species
}
\details{
Returns ggplot objects
}
