% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{plotSubVar}
\alias{plotSubVar}
\title{Plot substance variations}
\usage{
plotSubVar(simlist, metsel)
}
\arguments{
\item{simlist}{A list of simulations (eval objects).}

\item{metsel}{A vector with the name of exchange reactions of interest}
}
\description{
The function \code{plotSubVar} takes a list of simulations and returns a barplot with most varying substances
}
