% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\name{rmSubs}
\alias{rmSubs}
\alias{rmSubs,Arena-method}
\title{Remove substances}
\usage{
rmSubs(object, mediac)

\S4method{rmSubs}{Arena}(object, mediac)
}
\arguments{
\item{object}{An object of class Arena.}

\item{mediac}{A character vector giving the names of substances, which should be added to the environment (the default takes all possible substances).}
}
\description{
The generic function \code{rmSubs} removes all amounts of substances available in the arena for given compounds.
}
