% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organism.R
\name{setKinetics}
\alias{setKinetics}
\alias{setKinetics,Organism-method}
\title{Function to set Michaelis-Menten kinetics for uptake of a substance}
\usage{
setKinetics(object, exchangeR, Km, vmax)

\S4method{setKinetics}{Organism}(object, exchangeR, Km, vmax)
}
\arguments{
\item{object}{An object of class Organisms.}

\item{exchangeR}{Name of an exchange reaction}

\item{Km}{Parameter Michaelis-Menten-Kinetics (in mM)}

\item{vmax}{Parameter Michaelis-Menten-Kinetics (in mmol/(g*h))}
}
\description{
The generic function \code{setKinetics} provides kinetics for exchange reactions.
}
