% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sihumi_test}
\alias{sihumi_test}
\title{Multi-species test data set}
\format{An object of class \code{eval}}
\usage{
data(sihumi_test)
}
\description{
Data contains results of a simulation with 8 gut bacteria for 10 time steps
}
\keyword{datasets}
