% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Arena.R
\name{simulate}
\alias{simulate}
\title{Main function for simulating all processes in the environment}
\usage{
simulate(object, time)
}
\arguments{
\item{object}{An object of class Arena or Eval.}

\item{time}{A number giving the number of iterations to perform for the simulation}
}
\value{
Returns an object of class \code{Eval} which can be used for subsequent analysis steps.
}
\description{
The generic function \code{simulate} for a simple simulation of the environment.
}
\details{
The returned object itself can be used for a subsequent simulation, due to the inheritance between \code{Eval} and \code{Arena}.
}
\examples{
\dontrun{
ecore <- model #get Escherichia coli core metabolic model
bac <- Bac(ecore,deathrate=0.05,duplirate=0.5,
           growthlimit=0.05,growtype="exponential") #initialize a bacterium
arena <- Arena(20,20) #initialize the environment
addOrg(arena,bac,amount=10) #add 10 organisms
addSubs(arena,40) #add all possible substances
eval <- simulate(arena,10)
}
}
\seealso{
\code{\link{Arena-class}} and \code{\link{Eval-class}}
}

