% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Arena.R
\name{stirEnv}
\alias{stirEnv}
\title{Function for stirring/mixing the complete evironment}
\usage{
stirEnv(object, sublb)
}
\arguments{
\item{object}{An object of class Arena.}

\item{sublb}{A matrix with the substrate concentration for every individual in the environment based on their x and y position.}
}
\value{
Returns the substrate concentration for every individual in the environment with substrates as well as x and y positions as columns and rows for each organism.
}
\description{
The generic function \code{stirEnv} simulates the event of mixing all substrates and organisms in the environment.
}
\details{
The stirring is implemented as a random permutation of organism positions and the equalization of of all substrate concentrations.
}
\examples{
\dontrun{
ecore <- model #get Escherichia coli core metabolic model
bac <- Bac(ecore,deathrate=0.05,duplirate=0.5,
           growthlimit=0.05,growtype="exponential") #initialize a bacterium
arena <- Arena(20,20) #initialize the environment
addOrg(arena,bac,amount=10) #add 10 organisms
addSubs(arena,40) #add all possible substances
sublb <- getSublb(arena)
stirEnv(arena,sublb)
}
}
\seealso{
\code{\link{Arena-class}} and \code{\link{getSublb}}
}

