% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Substance.R
\docType{class}
\name{Substance-class}
\alias{Substance-class}
\title{Structure of the S4 class "Substance"}
\description{
Structure of the S4 class \code{Substance} representing substances in the environment which can be produced or consumed.
}
\section{Slots}{

\describe{
\item{\code{smax}}{A number representing the start concentration of the substance for each grid cell in the environment.}

\item{\code{diffmat}}{A sparse matrix containing all concentrations of the substance in the environment.}

\item{\code{name}}{A character vector representing the name of the substance.}

\item{\code{difunc}}{A character vector ("cpp" or "r") describing the function for diffusion.}

\item{\code{difspeed}}{A number indicating the diffusion speed (given by number of cells per iteration).}
}}

