% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Substance.R
\name{Substance-constructor}
\alias{Substance}
\alias{Substance-constructor}
\title{Constructor of the S4 class \code{Substance}}
\usage{
Substance(n, m, smax, gridgeometry, difspeed = 6.7e-06, ...)
}
\arguments{
\item{n}{A number giving the horizontal size of the environment.}

\item{m}{A number giving the vertical size of the environment.}

\item{smax}{A number representing the start concentration of the substance for each grid cell in the environment.}

\item{gridgeometry}{A list containing grid geometry parameter}

\item{difspeed}{A number indicating the diffusion speed (given by cm^2/s).}

\item{...}{Arguments of \code{\link{Substance-class}}}
}
\value{
Object of class \code{Substance}
}
\description{
The constructor to get a new object of class \code{Substance}
}

