% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\name{unit_conversion}
\alias{unit_conversion}
\alias{unit_conversion,Arena-method}
\title{Function for unit conversion}
\usage{
unit_conversion(object, unit)

\S4method{unit_conversion}{Arena}(object, unit)
}
\arguments{
\item{object}{An object of class Arena or Eval.}

\item{unit}{Unit to be converted to fmol/cell}
}
\value{
Conversion factor
}
\description{
The generic function \code{unit_conversion} converts units for e.g. substance concentratins
}
