% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\name{addEssentialMed}
\alias{addEssentialMed}
\alias{addEssentialMed,Arena-method}
\title{Add minimal medium of an organism to arena.}
\usage{
addEssentialMed(object, org, only_return = FALSE, limit = 10)

\S4method{addEssentialMed}{Arena}(object, org, only_return = FALSE, limit = 10)
}
\arguments{
\item{object}{An object of class Arena.}

\item{org}{An object of class Organism}

\item{only_return}{Set true if essential metabolites should only be returned but not added to arena}

\item{limit}{A metabolite is considered as essential if its remove whould decrease biomass growth below limit (between 0,100; default 10\%)}
}
\description{
The generic function \code{addEssentialMed} uses flux variability analysis to determine a essential growth medium  components.
}
