% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Arena.R
\name{addOrg}
\alias{addOrg}
\title{Function with to add individuals to the arena}
\usage{
addOrg(object, specI, amount, x = NULL, y = NULL, growth = 1, ...)
}
\arguments{
\item{object}{An object of class Arena}

\item{specI}{An object of class Organism}

\item{amount}{Number of organisms to add}

\item{x}{x positions of individuals on the grid}

\item{y}{y positions of individuals on the grid}

\item{growth}{Starting biomass of organisms}
}
\description{
Function with to add individuals to the arena
}
\examples{
NULL
}

