% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Organism.R
\docType{class}
\name{Bac-class}
\alias{Bac-class}
\title{Structure of the S4 class \code{Bac} inheriting from class \code{\link{Organism}} representing bacterial cells.}
\description{
Structure of the S4 class \code{Bac} inheriting from class \code{\link{Organism}} representing bacterial cells.
}
\section{Slots}{

\describe{
\item{\code{speed}}{A integer vector representing the speed by which bacterium is moving (given by cell per iteration).}

\item{\code{budge}}{A boolean vector indicating if budging (bacteria in the souronding area are pushed away) should be implemented.}

\item{\code{chem}}{A character vector indicating name of substance which is the chemotaxis attractant. Empty character vector if no chemotaxis.}
}}

