% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organism.R
\name{Bac-Constructor}
\alias{Bac-Constructor}
\alias{Bac}
\title{Constructor of the S4 class \code{\link{Bac-class}}}
\usage{
Bac(model, chem = "", ...)
}
\arguments{
\item{model}{model}

\item{chem}{A character vector indicating name of substance which is the chemotaxis attractant. Empty character vector if no chemotaxis.}

\item{...}{Arguments of \code{\link{Organism}}}
}
\value{
Object of class \code{\link{Bac-class}}
}
\description{
Constructor of the S4 class \code{\link{Bac-class}}
}
