% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organism.R
\docType{class}
\name{Bac-class}
\alias{Bac-class}
\title{Structure of the S4 class "Bac"}
\description{
Structure of the S4 class \code{Bac} inheriting from class \code{\link{Organism-class}} representing bacterial cells.
}
\section{Slots}{

\describe{
\item{\code{deathrate}}{A numeric value giving the factor by which the biomass should be reduced in every iteration if no growth is possible (default (E.coli): 0.21 pg)}

\item{\code{minweight}}{A numeric value giving the growth limit at which the organism dies. (default (E.coli): 0.083 pg)}

\item{\code{cellarea}}{A numeric value indicating the surface that one organism occupies (default (E.coli): 4.42 mu_m^2)}

\item{\code{maxweight}}{A numeric value giving the maximal dry weight of single organism (default (E.coli): 1.172 pg)}

\item{\code{chem}}{A character vector indicating name of substance which is the chemotaxis attractant. Empty character vector if no chemotaxis.}
}}

