% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\name{diffuse}
\alias{diffuse}
\alias{diffuse,Arena-method}
\title{Function for diffusion}
\usage{
diffuse(object, lrw, sublb, verbose = TRUE)

\S4method{diffuse}{Arena}(object, lrw, sublb, verbose = TRUE)
}
\arguments{
\item{object}{An object of class Arena.}

\item{lrw}{A numeric value needed by solver to estimate array size (by default lwr is estimated in the simEnv() by the function estimate_lrw())}

\item{sublb}{A matrix with the substrate concentration for every individual in the environment based on their x and y position.}

\item{verbose}{Set to false if no status messages should be printed.}
}
\description{
The generic function \code{diffuse} computes the media distribution via diffusion
}
