% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Substance.R
\name{diffusePDE}
\alias{diffusePDE}
\alias{diffusePDE,Substance-method}
\title{Function for diffusion of the Substance matrix}
\usage{
diffusePDE(object, init_mat, gridgeometry, lrw = NULL, tstep)

\S4method{diffusePDE}{Substance}(object, init_mat, gridgeometry, lrw = NULL, tstep)
}
\arguments{
\item{object}{An object of class Substance.}

\item{init_mat}{A matrix with values to be used by the diffusion.}

\item{gridgeometry}{A list specifying the geometry of the Arena}

\item{lrw}{A numeric value needed by solver to estimate array size (by default lwr is estimated in simEnv() by the function estimate_lrw())}

\item{tstep}{A numeric value giving the time step of integration}
}
\description{
The generic function \code{diffusePDE} implements the diffusion by the solving diffusion equation.
}
\details{
Partial differential equation is solved to model 2d diffusion process in the arena.
}
\seealso{
\code{\link{Substance-class}} and \code{\link{diffuseR}}
}
