% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\docType{class}
\name{Eval-class}
\alias{Eval-class}
\title{Structure of the S4 class "Eval"}
\description{
Structure of the S4 class \code{Eval} inheriting from class \code{\link{Arena-class}} for the analysis of simulations.
}
\section{Slots}{

\describe{
\item{\code{medlist}}{A list of compressed medium concentrations (only changes of concentrations are stored) per time step.}

\item{\code{simlist}}{A list of the organism features per time step.}

\item{\code{mfluxlist}}{A list of containing highly used metabolic reactions per time step.}

\item{\code{shadowlist}}{A list of containing shadow prices per time step.}

\item{\code{subchange}}{A vector of all substrates with numbers indicating the degree of change in the overall simulation.}

\item{\code{exchangeslist}}{A list of containing exchanges per time step.}
}}

