% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\name{findFeeding}
\alias{findFeeding}
\alias{findFeeding,Eval-method}
\title{Function for investigation of feeding between phenotypes}
\usage{
findFeeding(
  object,
  dict = NULL,
  tcut = 5,
  scut = NULL,
  org_dict = NULL,
  legendpos = "topleft",
  lwd = 1
)

\S4method{findFeeding}{Eval}(
  object,
  dict = NULL,
  tcut = 5,
  scut = NULL,
  org_dict = NULL,
  legendpos = "topleft",
  lwd = 1
)
}
\arguments{
\item{object}{An object of class Eval.}

\item{dict}{List defining new substance names. List entries are intepreted as old names and the list names as the new ones.}

\item{tcut}{Integer giving the minimal mutual occurence ot be considered (dismiss very seldom feedings)}

\item{scut}{substance names which should be ignored}

\item{org_dict}{A named list/vector with names that should replace (eg. unreadable) IDs}

\item{legendpos}{A character variable declaring the position of the legend}

\item{lwd}{Line thickness scale in graph}
}
\value{
Graph (igraph)
}
\description{
The generic function \code{findFeeding}
}
