% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\name{findFeeding3}
\alias{findFeeding3}
\alias{findFeeding3,Eval-method}
\title{Function for investigation of feeding between phenotypes}
\usage{
findFeeding3(object, time, mets, plot = TRUE, cutoff = 1e-06)

\S4method{findFeeding3}{Eval}(object, time, mets, plot = TRUE, cutoff = 1e-06)
}
\arguments{
\item{object}{An object of class Eval.}

\item{time}{A numeric vector giving the simulation steps which should be plotted.}

\item{mets}{Character vector of substance names which should be considered}

\item{plot}{Should the graph also be plotted?}

\item{cutoff}{Accuracy of crossfeeding interaction (minimal flux to be considered)}
}
\value{
Graph (igraph)
}
\description{
The generic function \code{findFeeding3}
}
