% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\name{getSubHist}
\alias{getSubHist}
\alias{getSubHist,Eval-method}
\title{Function to get timeline of a substance}
\usage{
getSubHist(object, sub, unit = "fmol/cell")

\S4method{getSubHist}{Eval}(object, sub, unit = "fmol/cell")
}
\arguments{
\item{object}{An object of class Eval.}

\item{sub}{Vector with substances.}

\item{unit}{Unit to be used}
}
\description{
The generic function \code{getSubHist} returns list with amount of substance for each timestep
}
