% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Bac.R
\name{growth}
\alias{growth}
\title{Function with the growth model of a bac (biomass growth, replication, death)}
\usage{
growth(object, population, j)
}
\arguments{
\item{object}{An object of class Bac}

\item{population}{An object of class Arena}

\item{j}{The number of the iteration of interest}
}
\value{
Boolean variable of the \code{j}th individual indicating if individual died.
}
\description{
Function with the growth model of a bac (biomass growth, replication, death)
}
\examples{
NULL
}

