% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organism.R
\name{growth_par}
\alias{growth_par}
\alias{growth_par,Bac-method}
\title{Function implementing a growth model of a bacterium}
\usage{
growth_par(object, population, j, fbasol, tstep)

\S4method{growth_par}{Bac}(object, population, j, fbasol, tstep)
}
\arguments{
\item{object}{An object of class Bac.}

\item{population}{An object of class Arena.}

\item{j}{The index of the organism of interest in orgdat.}

\item{fbasol}{Problem object according to the constraints and then solved with \code{optimizeProb}.}

\item{tstep}{A number giving the time intervals for each simulation step.}
}
\value{
A list
}
\description{
The generic function \code{growth_par} implements different growth models for an object of class Bac.
}
