% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{plotAbundance}
\alias{plotAbundance}
\title{Plot abundances of species}
\usage{
plotAbundance(
  simlist,
  time = c(NULL, NULL),
  col = colpal3,
  return_dat = F,
  use_biomass = F
)
}
\arguments{
\item{simlist}{A list of simulations (eval objects).}

\item{time}{A vector with start and end time to be considered (default: total time)}

\item{col}{Vector with color that should be used}

\item{return_dat}{Should plain text mean abundances be returned? (default false)}

\item{use_biomass}{If enabled then biomass is used instead of cell number}
}
\description{
The function \code{plotAbundance} takes a list of simulations and return a boxplot with species abundances
}
