% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{plotGrowthCurve}
\alias{plotGrowthCurve}
\title{Plot growth curve for several simulations}
\usage{
plotGrowthCurve(simlist, bcol = colpal3, time = c(NULL, NULL), title = "",
  size = 1)
}
\arguments{
\item{simlist}{A list of simulations (eval objects).}

\item{bcol}{Vector with color that should be used}

\item{time}{Vector with two entries defining start and end time}
}
\description{
The function \code{plotGrowthCurve} takes a list of simulations and plots the time course of species with standard deviation.
}

