% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\docType{methods}
\name{plotSpecActivity}
\alias{plotSpecActivity}
\alias{plotSpecActivity,Eval-method}
\title{Function to plot substance usage for every species}
\usage{
plotSpecActivity(object, subs = list(), var_nr = 10)

\S4method{plotSpecActivity}{Eval}(object, subs = list(), var_nr = 10)
}
\arguments{
\item{object}{An object of class Eval.}

\item{subs}{List of substance names}

\item{var_nr}{Number of most varying substances to be used (if subs is not specified)}
}
\description{
The generic function \code{plotSpecActivity} displays the input/output substances with the highest variance (could also be defiened manually) for all species
}
\details{
Returns ggplot objects
}

