% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{plotSpecActivity}
\alias{plotSpecActivity}
\title{Function to plot substance usage for every species}
\usage{
plotSpecActivity(simlist, subs = list(), var_nr = 10, spec_list = NULL)
}
\arguments{
\item{simlist}{An object of class Eval or a list with objects of class Eval.}

\item{subs}{List of substance names}

\item{var_nr}{Number of most varying substances to be used (if subs is not specified)}

\item{spec_list}{List of species names to be considered (default all)}
}
\description{
The generic function \code{plotSpecActivity} displays the input/output substances with the highest variance (could also be defiened manually) for all species
}
\details{
Returns ggplot objects
}

