% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stuff.R
\name{plotSpecActivity}
\alias{plotSpecActivity}
\title{Function to plot substance usage for every species}
\usage{
plotSpecActivity(
  simlist,
  subs = list(),
  var_nr = 10,
  spec_list = NULL,
  ret_data = FALSE,
  useNames = FALSE,
  rm_unused = TRUE,
  cutoff = 1e-06
)
}
\arguments{
\item{simlist}{An object of class Eval or a list with objects of class Eval.}

\item{subs}{List of substance names}

\item{var_nr}{Number of most varying substances to be used (if subs is not specified)}

\item{spec_list}{List of species names to be considered (default all)}

\item{ret_data}{Set true if data should be returned}

\item{useNames}{Use substance names instead of ids}

\item{rm_unused}{Remove substances which do not change from plot}

\item{cutoff}{Minimum valu for fluxes to be considered}
}
\description{
The generic function \code{plotSpecActivity} displays the input/output substances with the highest variance (could also be defiened manually) for all species
}
\details{
Returns ggplot objects
}
